rename list and request button folders
