+++ feature-libs/quote/components/list/quote-list-component.service.spec.ts
