+++ feature-libs/quote/components/list/quote-list.component.ts
