+++ feature-libs/quote/components/list/quote-list.module.ts
