+++ feature-libs/quote/components/quote-components.module.ts
@@ -13,8 +13,8 @@ import { QuoteActionsConfirmDialogModule } from './actions/confirm-dialog/quote-
 import { QuoteActionsLinkModule } from './actions/link/quote-actions-link.module';
 import { defaultQuoteUIConfig } from './config/default-quote-ui.config';
 import { QuoteSellerEditModule } from './header/seller-edit/quote-seller-edit.module';
-import { QuoteListModule } from './quote-list/quote-list.module';
-import { QuoteRequestButtonModule } from './quote-request-button/quote-request-button.module';
+import { QuoteListModule } from './list/quote-list.module';
+import { QuoteRequestButtonModule } from './request-button/quote-request-button.module';
 import { QuoteHeaderBuyerEditModule } from './header/buyer-edit/quote-header-buyer-edit.module';
 import { QuoteHeaderOverviewModule } from './header/overview/quote-header-overview.module';
 import { QuoteCommentsModule } from './comments/quote-comments.module';
